Aggiunto controllo di integrità in fase di inserimento di una nuova persona riguardo ai vacation periods

+++ app/models/Contract.java
@@ -38,23 +38,23 @@ import play.db.jpa.Model;
 @Entity
 @Table(name="contracts")
 public class Contract extends Model {
-	
+
 	private static final long serialVersionUID = -4472102414284745470L;
-	
+
 	@ManyToOne(fetch=FetchType.LAZY)
 	@JoinColumn(name="person_id")
 	public Person person;
-	
+
 	/**
 	 * relazione con la tabella di vacation_code
 	 */
 	@OneToMany(mappedBy="contract", fetch=FetchType.LAZY, cascade = CascadeType.REMOVE)
 	public List<VacationPeriod> vacationPeriods;
-	
+
 	@Type(type="org.joda.time.contrib.hibernate.PersistentLocalDate")
 	@Column(name="begin_contract")
 	public LocalDate beginContract;
-	
+
 	@Type(type="org.joda.time.contrib.hibernate.PersistentLocalDate")
 	@Column(name="expire_contract")
 	public LocalDate expireContract;
@@ -65,7 +65,7 @@ public class Contract extends Model {
 	@Type(type="org.joda.time.contrib.hibernate.PersistentLocalDate")
 	@Column(name="end_contract")
 	public LocalDate endContract;
-	
+
 	//TODO eliminare e configurare yaml
 	public void setBeginContract(String date){
 		this.beginContract = new LocalDate(date);
@@ -78,7 +78,7 @@ public class Contract extends Model {
 	public void setExpireContract(String date){
 		this.expireContract = new LocalDate(date);
 	}
-	
+
 	/**
 	 * I contratti con onCertificate = true sono quelli dei dipendenti CNR e 
 	 * corrispondono a quelli con l'obbligo dell'attestato di presenza 
@@ -91,15 +91,15 @@ public class Contract extends Model {
 	public boolean isValidContract(){
 		LocalDate date = new LocalDate();
 		return endContract==null && beginContract.isBefore(date) && expireContract.isAfter(date);
-					 
+
 	}
-	
+
 	@Override
 	public String toString() {
 		return String.format("Contract[%d] - person.id = %d, beginContract = %s, expireContract = %s, endContract = %s",
-			id, person.id, beginContract, expireContract, endContract);
+				id, person.id, beginContract, expireContract, endContract);
 	}
-	
+
 	/**
 	 * @param contract
 	 * @return il vacation period associato al contratto con al suo interno la data di oggi
@@ -109,15 +109,15 @@ public class Contract extends Model {
 		List<VacationPeriod> vpList = this.getContractVacationPeriods();
 		for(VacationPeriod vp : vpList)
 		{
-			
+
 			LocalDate now = new LocalDate();
-			
+
 			if(DateUtility.isDateIntoInterval(now, new DateInterval(vp.beginFrom, vp.endTo)))
 				return vp;
 		}
 		return null;
 	}
-	
+
 	/**
 	 * @param contract
 	 * @return i vacation period associati al contratto, ordinati in ordine crescente per data inizio
@@ -127,34 +127,36 @@ public class Contract extends Model {
 	{
 		//vacation period piu' recente per la persona
 		List<VacationPeriod> vpList = VacationPeriod.find(  "SELECT vp "
-													+ "FROM VacationPeriod vp "
-													+ "WHERE vp.contract = ? "
-													+ "ORDER BY vp.beginFrom",
-													this).fetch();
-		
+				+ "FROM VacationPeriod vp "
+				+ "WHERE vp.contract = ? "
+				+ "ORDER BY vp.beginFrom",
+				this).fetch();
+
 		//se il piano ferie associato al contratto non esiste 
 		if(vpList==null)
 		{
 			Logger.debug("CurrentPersonVacationPeriod: il vacation period è inesistente");
 			return null;
 		}
-		
-		
+
+
 		return vpList;
 	}
-	
+
 	/**
 	 * Costruisce la struttura valida dei vacation period associati al contratto.
 	 */
 	public void setVacationPeriods(){
-		
-		
+
+
 		//Distruggo i vacation period precedenti
-		for(VacationPeriod vp : this.vacationPeriods)
-		{
-			vp.delete();
+		if(this.vacationPeriods != null){
+			for(VacationPeriod vp : this.vacationPeriods)
+			{
+				vp.delete();
+			}
 		}
-		
+
 		//Contratto terminato nessun vacation period
 		if(this.endContract!=null)
 		{
@@ -165,7 +167,7 @@ public class Contract extends Model {
 		//Tempo indeterminato, creo due vacatio 3 anni più infinito
 		if(this.expireContract == null)
 		{
-			
+
 			VacationPeriod first = new VacationPeriod();
 			first.beginFrom = this.beginContract;
 			first.endTo = this.beginContract.plusYears(3).minusDays(1);
